requirements.txt
python-telegram-bot[rate-limiter]==20.1
python-dotenv==0.21.0
openai==0.28.1